Added 'next' to get next state
